+++ examples/distributed-computing/workitem.py
@@ -8,3 +8,12 @@ class Workitem(object):
 
     def __str__(self):
         return "<Workitem id=%s>" % str(self.itemId)
+
+    @staticmethod
+    def from_dict(classname, d):
+        """this method is used to deserialize a workitem from Pyro"""
+        assert classname == "workitem.Workitem"
+        w = Workitem(d["itemId"], d["data"])
+        w.result = d["result"]
+        w.processedBy = d["processedBy"]
+        return w